requirements.txt
#Github:Rainilyl
pyyaml==6.0
paramiko==2.10.4
gitpython==3.1.20